+++ src/euroeval/data_models.py
@@ -10,11 +10,10 @@ from dataclasses import dataclass, field
 import pydantic
 import torch
 
-from euroeval.utils import get_package_version
-
 from .enums import Device, InferenceBackend, ModelType, TaskGroup
 from .templates import LANG_TO_OR, TEMPLATES_DICT
 from .types import ScoreDict
+from .utils import get_package_version
 
 
 @dataclass